Support fractional monitor speeds and update Constructor mode documentation
Refactors the monitor speed system from integer to floating-point values to support a 12.5 mm/s speed option in Teaching mode. Additionally, performs extensive updates to the project's technical documentation and data parsing logic to reflect the transition from "Editor" to "Constructor" mode, including support for ECG landmark markers.

- **Monitor & UI:**
    - Update `MonitorModeModel` and `MonitorViewModel` to treat `speed` as a `Float`.
    - Enhance `MonitorControlPanel` and `ConstructorControlPanel` to format speed strings dynamically (e.g., "12.5" or "25") and allow decimal input.
    - Add 12.5 mm/s to the available speeds list specifically for `OperatingMode.Teaching`.
    - Update string resources across all locales to use string formatting (`%1$s`) for speeds instead of integer formatting (`%1$d`).

- **Data & Persistence:**
    - Update `DataSourcePrefs` to store monitor speed as a `Float`.
    - Implement robust type-conversion helpers in `DataSourcePrefs` (e.g., `getFloat`, `getInt`) to maintain backward compatibility with existing integer-based preference keys.
    - Refactor `PathologyParser` and the data structure specification to support the optional `markers:` header field for significant ECG points.

- **Constructor Mode:**
    - Update `ConstructorViewModel` to verify the success of the `writePathology` operation before clearing dirty states.
    - Rename and document various "Editor" components to "Constructor" to align with the refined domain model.

- **Documentation:**
    - Update `architecture.md` to detail the current MVVM-with-Compose structure, including the `SignificantPointOverlay` and synchronized scrolling logic.
    - Revise `ecg-rendering-pipeline.md` to describe the five-stage rendering process, "Sweep" display mode for blank grids, and interval bracket calculations.
    - Update `tcp-protocol.md` to clarify the connection lifecycle and socket management implemented in `AppViewModel`.

+++ app/src/main/java/com/example/cardiosimulator/domain/MonitorModeModel.kt
@@ -20,7 +20,7 @@ data class MonitorModeModel(
     val count: Int = 1,
     val gridScheme: GridScheme = GridScheme.Pink,
     val seriesScheme: SeriesScheme = SeriesScheme.OneColumn,
-    val speed: Int = 25,
+    val speed: Float = 25f,
     val scale: Float = 1f,
     val displayScale: Float = 0.4f,
     val calibration: EcgCalibration = EcgCalibration(),
